refactor: 테스트 코드 baseUrl 정리 및 패키지 분류
