+++ src/test/java/gift/ControllerTest/MemberControllerTest.java
@@ -1,11 +1,9 @@
-package gift;
+package gift.ControllerTest;
 
+import gift.Application;
 import gift.dto.MemberRequestDto;
 import gift.dto.MemberResponseDto;
-import org.junit.jupiter.api.MethodOrderer;
-import org.junit.jupiter.api.Order;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
+import org.junit.jupiter.api.*;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
@@ -26,12 +24,19 @@ public class MemberControllerTest {
 
     private RestClient client = RestClient.builder().build();
 
+    private String baseUrl;
+
+    @BeforeEach
+    void setUp() {
+        baseUrl = "http://localhost:" + port + "/api/members";
+    }
+
     @Test
     @Order(1)
     void 회원가입_잘못된_이메일_입력_테스트(){
         System.out.println("Member Register Not Valid Email test");
         MemberRequestDto requestDto = new MemberRequestDto("qwertypusan.ac.kr", "12345678");
-        var url = "http://localhost:" + port + "/api/members/membership";
+        String url = baseUrl + "/membership";
         assertThatExceptionOfType(HttpClientErrorException.BadRequest.class)
                 .isThrownBy(
                         () ->
@@ -49,7 +54,7 @@ public class MemberControllerTest {
     void 회원가입_잘못된_비밀번호_입력_테스트(){
         System.out.println("Member Register Not Valid Email test");
         MemberRequestDto requestDto = new MemberRequestDto("abcd@pusan.ac.kr", "1234");
-        var url = "http://localhost:" + port + "/api/members/membership";
+        String url = baseUrl + "/membership";
         assertThatExceptionOfType(HttpClientErrorException.BadRequest.class)
                 .isThrownBy(
                         () ->
@@ -67,7 +72,7 @@ public class MemberControllerTest {
     void 회원가입_중복_이메일_테스트(){
         System.out.println("Elready Exist Email Register test");
         MemberRequestDto requestDto = new MemberRequestDto("abc@pusan.ac.kr", "12345678");
-        var url = "http://localhost:" + port + "/api/members/membership";
+        String url = baseUrl + "/membership";
         assertThatExceptionOfType(HttpClientErrorException.BadRequest.class)
                 .isThrownBy(
                         () -> client.post()
@@ -84,7 +89,7 @@ public class MemberControllerTest {
     void 회원가입_정상_테스트(){
         System.out.println("Member Register Success test");
         MemberRequestDto requestDto = new MemberRequestDto("zxc@pusan.ac.kr", "12345678");
-        var url = "http://localhost:" + port + "/api/members/membership";
+        String url = baseUrl + "/membership";
         var response = client.post()
                 .uri(url)
                 .contentType(MediaType.APPLICATION_JSON)
@@ -100,7 +105,7 @@ public class MemberControllerTest {
     void 로그인_잘못된_이메일_입력_테스트(){
         System.out.println("Member Login Not Valid Email test");
         MemberRequestDto requestDto = new MemberRequestDto("qwertypusan.ac.kr", "12345678");
-        var url = "http://localhost:" + port + "/api/members/login";
+        String url = baseUrl + "/login";
         assertThatExceptionOfType(HttpClientErrorException.BadRequest.class)
                 .isThrownBy(
                         () ->
@@ -118,7 +123,7 @@ public class MemberControllerTest {
     void 로그인_잘못된_비밀번호_입력_테스트(){
         System.out.println("Member Login Not Valid Email test");
         MemberRequestDto requestDto = new MemberRequestDto("abcd@pusan.ac.kr", "1234");
-        var url = "http://localhost:" + port + "/api/members/login";
+        String url = baseUrl + "/login";
         assertThatExceptionOfType(HttpClientErrorException.BadRequest.class)
                 .isThrownBy(
                         () ->
@@ -136,7 +141,7 @@ public class MemberControllerTest {
     void 로그인_없는_이메일_시도_테스트(){
         System.out.println("Member Login Not register Email test");
         MemberRequestDto requestDto = new MemberRequestDto("qwerty@pusan.ac.kr", "12345678");
-        var url = "http://localhost:" + port + "/api/members/login";
+        String url = baseUrl + "/login";
         assertThatExceptionOfType(HttpClientErrorException.NotFound.class)
                 .isThrownBy(
                         () ->
@@ -154,7 +159,7 @@ public class MemberControllerTest {
     void 로그인_틀린_비밀번호_테스트(){
         System.out.println("Member Login Invalid password test");
         MemberRequestDto requestDto = new MemberRequestDto("abcd@pusan.ac.kr", "12345678888");
-        var url = "http://localhost:" + port + "/api/members/login";
+        String url = baseUrl + "/login";
         assertThatExceptionOfType(HttpClientErrorException.Forbidden.class)
                 .isThrownBy(
                         () -> client.post()
@@ -171,7 +176,7 @@ public class MemberControllerTest {
     void 로그인_정상_테스트(){
         System.out.println("Member Login success test");
         MemberRequestDto requestDto = new MemberRequestDto("abcd@pusan.ac.kr", "12345678");
-        var url = "http://localhost:" + port + "/api/members/login";
+        String url = baseUrl + "/login";
         var response = client.post()
                 .uri(url)
                 .contentType(MediaType.APPLICATION_JSON)
